refactor: move Data structs from lib::accounting to wire crate
