+++ lib/src/accounting/server_state.rs
@@ -2,11 +2,13 @@ use crate::common::{request, request_bare, SerializableNone};
 use crate::error::ApiError;
 use anyhow::Context;
 use chrono::{DateTime, FixedOffset};
-use lrzcc_wire::accounting::{ServerState, ServerStateImport};
+use lrzcc_wire::accounting::{
+    ServerState, ServerStateCreateData, ServerStateImport,
+    ServerStateModifyData,
+};
 use reqwest::blocking::Client;
 use reqwest::Url;
 use reqwest::{Method, StatusCode};
-use serde::Serialize;
 use std::rc::Rc;
 
 pub struct ServerStateApi {
@@ -84,40 +86,6 @@ impl ServerStateListRequest {
     }
 }
 
-#[derive(Clone, Debug, Serialize)]
-struct ServerStateCreateData {
-    begin: DateTime<FixedOffset>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    end: Option<DateTime<FixedOffset>>,
-    instance_id: String, // UUIDv4
-    instance_name: String,
-    flavor: u32,
-    // TODO we need an enum here
-    status: String,
-    user: u32,
-}
-
-impl ServerStateCreateData {
-    fn new(
-        begin: DateTime<FixedOffset>,
-        instance_id: String, // UUIDv4
-        instance_name: String,
-        flavor: u32,
-        status: String,
-        user: u32,
-    ) -> Self {
-        Self {
-            begin,
-            end: None,
-            instance_id,
-            instance_name,
-            flavor,
-            status,
-            user,
-        }
-    }
-}
-
 pub struct ServerStateCreateRequest {
     url: String,
     client: Rc<Client>,
@@ -164,42 +132,6 @@ impl ServerStateCreateRequest {
             Some(&self.data),
             StatusCode::CREATED,
         )
-    }
-}
-
-#[derive(Clone, Debug, Serialize)]
-struct ServerStateModifyData {
-    id: u32,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    begin: Option<DateTime<FixedOffset>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    end: Option<DateTime<FixedOffset>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    instance_id: Option<String>, // UUIDv4
-    #[serde(skip_serializing_if = "Option::is_none")]
-    instance_name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    flavor: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    // TODO we need an enum here
-    status: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    user: Option<u32>,
-}
-
-impl ServerStateModifyData {
-    fn new(id: u32) -> Self {
-        Self {
-            id,
-            begin: None,
-            end: None,
-            instance_id: None,
-            instance_name: None,
-            flavor: None,
-            status: None,
-            user: None,
-        }
     }
 }
